Moved validations method into the cron_common partial. Removed cron validations helper files. Added another common partial for cron and cron_d resource properties. Added chef_client_cron resource into the resource/cron directory. Removed unnecessary validation methods. Added rspec for validation methods.
