+++ lib/chef/resources.rb
@@ -27,7 +27,7 @@ require_relative "resource/batch"
 require_relative "resource/breakpoint"
 require_relative "resource/build_essential"
 require_relative "resource/cookbook_file"
-require_relative "resource/chef_client_cron"
+require_relative "resource/cron/chef_client_cron"
 require_relative "resource/chef_client_scheduled_task"
 require_relative "resource/chef_client_systemd_timer"
 require_relative "resource/chef_gem"
